feat(action-center): 2-col grid layout + extract TravelCard from InvitationCard
- Rename InvitationCard section label "INVITATION" → "CREW EMAIL"
- Rewrite helper text to mention future blasts
- Extract travel toggle + form into standalone TravelCard in ActionCenter
- Going stage renders as CSS grid (auto-fit, minmax 280px) — 2 cols on desktop, stacks on mobile
- TravelCard returns null for non-owners when travel is disabled (grid auto-collapses)
